Split up GeoCoordCalculator and popgen utils, removed AgeDistribution
