+++ src/main/cpp/popgen/PopulationGenerator.cpp
@@ -1,5 +1,4 @@
 #include "PopulationGenerator.h"
-#include "AgeDistribution.h"
 #include "FamilyParser.h"
 #include "util/InstallDirs.h"
 #include "util/TimeStamp.h"
